requirements.txt
alabaster==0.7.12
arabic-reshaper==2.1.3
asgiref==3.3.4
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
Django==3.2.4
django-phonenumber-field==5.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
docutils==0.17.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
imagesize==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
Pillow==8.2.0
Pygments==2.9.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
